+++ SocialCreditScoreBot2/Implementations/VaderSharp.cs
@@ -6,11 +6,11 @@ namespace SocialCreditScoreBot2.Implementations;
 public class VaderSharp : ISentimentAnalyzer {
     public Task<bool> Init() {return Task.FromResult(true);}
 
-    public Task<float> Analyse(string text) {
+    public Task<double> Analyse(string text) {
         SentimentIntensityAnalyzer analyzer = new SentimentIntensityAnalyzer();
         
         SentimentAnalysisResults results = analyzer.PolarityScores(text);
         
-        return Task.FromResult((float)results.Compound);
+        return Task.FromResult(results.Compound);
     }
 }
